refactor

+++ TAD/aluno.h
@@ -3,6 +3,6 @@ struct Aluno {
     char nome[50];
     char conceito;
     char disciplina[20];
-    int notas[3];
+    double notas[3];
     double media;
 };
